Version 0.2
Batch command parsing has been removed and an option to add genomes en masse has been added.
The Phylosift_PMPROK profile will now give errors if genomes do not contain any marker data.
The port on which the postgres database is running can be specified to the backend. And further to this, a --dev option has been added to the CLI to allow modifications to a separate database while developing.